// src/enums.ts
export enum Machine {
	START = "START",
	ARG_INT = "ARG_INT",
	ARG_FLOAT = "ARG_FLOAT",
	OPERATOR = "OPERATOR",
	PARENTHESES = "PARENTHESES",
	CALCULATE_RESULT = "CALCULATE_RESULT",
	CONTINUE_CALCULATION = "CONTINUE_CALCULATION"
}

export enum Keys {
	PARENTHESES_LEFT = "(",
	PARENTHESES_RIGHT = ")",
	DOT = ".",
	CLEAR = "C",
	CLEAR_ENTRY = "CE",
	DELETE = "Delete",
	MULTIPLY = "*",
	DIVIDE = "/",
	PLUS = "+",
	MINUS = "-",
	EQUALS = "=",
	PLUS_MINUS = "+/-",
	SQUARE_ROOT = "√",
	SQUARE = "x²",
	ZERO = "0",
	ONE = "1",
	TWO = "2",
	THREE = "3",
	FOUR = "4",
	FIVE = "5",
	SIX = "6",
	SEVEN = "7",
	EIGHT = "8",
	NINE = "9"
}